✨ Add Round API router
Identical features to the Stage router
